+++ workalendar/asia/israel.py
@@ -40,16 +40,21 @@ class Israel(Calendar):
                     days.append((current_date, "Sukkot"))
                 elif day == 22:
                     days.append((current_date, "Shmini Atzeres"))
-            elif month == 12:
-                leap = HebrewDate._is_leap(jewish_year)
+            elif month == 12 and not HebrewDate._is_leap(jewish_year):
                 if day == 14:
                     days.append((current_date, "Purim"))
-                if day == 15 and not leap:
+                if day == 15:
                     days.append((current_date, "Purim"))
+                    days.append((current_date, "Shushan Purim"))
+                elif day == 16:
+                    days.append((current_date, "Shushan Purim"))
             elif month == 13:
                 if day == 14:
                     days.append((current_date, "Purim"))
                 elif day == 15:
+                    days.append((current_date, "Purim"))
+                    days.append((current_date, "Shushan Purim"))
+                elif day == 16:
                     days.append((current_date, "Shushan Purim"))
             elif month == 1 and day in {15, 21}:
                 days.append((current_date, "Pesach"))